+++ packages/andromeda-data-storage/src/lib.rs
@@ -1,1 +1,4 @@
+pub mod boolean;
+pub mod counter;
 pub mod primitive;
+pub mod string_storage;